+++ lib/Model/Publisher/CsvStreamPublisher.php
@@ -4,7 +4,6 @@ namespace DTL\Extension\Fink\Model\Publisher;
 
 use DTL\Extension\Fink\Model\Publisher;
 use DTL\Extension\Fink\Model\Report;
-use RuntimeException;
 
 class CsvStreamPublisher implements Publisher
 {
@@ -23,16 +22,8 @@ class CsvStreamPublisher implements Publisher
      */
     private $stream;
 
-    public function __construct(string $path, bool $withHeaders)
+    public function __construct($stream, bool $withHeaders)
     {
-        $stream = fopen($path, 'w');
-        if (false === $stream) {
-            throw new RuntimeException(sprintf(
-                'Could not open stream for writing at path "%s"',
-                $path
-            ));
-        }
-
         $this->stream = $stream;
         $this->withHeaders = $withHeaders;
     }